Create separate row class for each assessment
+ The X2AssessmentRow was starting to fill up with if/then statements and other logic relating to different assessment types
+ Break each of these into their own class in order to handle different assessments without resorting to if blocks
