+++ app/importers/null_row.rb
@@ -1,0 +1,4 @@
+class NullRow
+  def save!; end
+end
+
