+++ app/importers/x2_assessment_importer.rb
@@ -19,7 +19,18 @@ class X2AssessmentImporter
 
   def import_row(row)
     return unless row[:assessment_test].match(WHITELIST)
-    X2AssessmentRow.build(row).save!
+
+    row[:assessment_growth] = nil if !/\D/.match(row[:assessment_growth]).nil?
+    row[:assessment_test] = "ACCESS" if row[:assessment_test] == "WIDA-ACCESS"
+
+    case row[:assessment_test]
+    when 'MCAS'
+      McasRow.build(row).save!
+    when 'ACCESS'
+      AccessRow.build(row).save!
+    when 'DIBELS'
+      DibelsRow.build(row).save!
+    end
   end
 
 end
